% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_life_expectancy.R
\name{calc_life_expectancy}
\alias{calc_life_expectancy}
\title{Calculate Life Expectancy}
\usage{
calc_life_expectancy(current_age, mode, dispersion, max_age = 120)
}
\arguments{
\item{current_age}{A numeric. Current age.}

\item{mode}{A numeric. Mode of the Gompertz distribution.}

\item{dispersion}{A numeric. Dispersion of the Gompertz distribution.}

\item{max_age}{A numeric. Maximum age. Defaults to 120.}
}
\value{
A numeric. Total life expectancy in years.
}
\description{
Calculate Life Expectancy
}
\examples{
calc_life_expectancy(
  current_age = 65, 
  mode        = 80, 
  dispersion  = 10
)
}
