% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_current_date}
\alias{get_current_date}
\title{Get current date}
\usage{
get_current_date()
}
\value{
A date.
}
\description{
If \code{R4GPF.current_date} option is not set, the current system date is used.
}
\examples{
get_current_date()
# Setting custom date using `R4GPF.current_date` option
options(R4GPF.current_date = as.Date("2023-01-01"))
get_current_date()
options(R4GPF.current_date = NULL) # Reset default date#' Working with cache

get_current_date()
}
