% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life_tables.R
\docType{data}
\name{life_tables}
\alias{life_tables}
\title{HMD life tables}
\format{
\subsection{\code{life_tables}}{

A data frame with 6 columns:
\describe{
\item{country}{Country name}
\item{sex}{Sex: "male", "female", "both"}
\item{year}{Year}
\item{age}{Age}
\item{mortality_rate}{Mortality rate}
\item{life_expectancy}{Life expectancy}
}
}
}
\source{
\url{https://www.mortality.org}
}
\usage{
life_tables
}
\description{
A data frame based on: HMD. Human Mortality Database. Max Planck Institute for Demographic Research (Germany), University of California, Berkeley (USA), and French Institute for Demographic Studies (France). Available at www.mortality.org.
}
\keyword{datasets}
