% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print_currency}
\alias{print_currency}
\title{Printing currency value}
\usage{
print_currency(
  x,
  suffix = "",
  big.mark = ",",
  accuracy = NULL,
  prefix = NULL,
  ...
)
}
\arguments{
\item{x}{A numeric vector}

\item{big.mark}{Character used between every 3 digits to separate thousands.}

\item{accuracy}{A number to round to. Use (e.g.) \code{0.01} to show 2 decimal
places of precision. If \code{NULL}, the default, uses a heuristic that should
ensure breaks have the minimum number of digits needed to show the
difference between adjacent values.

Applied to rescaled data.}

\item{prefix, suffix}{Symbols to display before and after value.}

\item{...}{Other arguments passed on to \code{\link[base:format]{base::format()}}.}
}
\description{
Printing currency value
}
