% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print_percent}
\alias{print_percent}
\title{Printing percent value}
\usage{
print_percent(x, accuracy = 0.1, ...)
}
\arguments{
\item{x}{A numeric vector}

\item{accuracy}{A number to round to. Use (e.g.) \code{0.01} to show 2 decimal
places of precision. If \code{NULL}, the default, uses a heuristic that should
ensure breaks have the minimum number of digits needed to show the
difference between adjacent values.

Applied to rescaled data.}

\item{...}{Other arguments passed on to \code{\link[base:format]{base::format()}}.}
}
\description{
Printing percent value
}
