% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_optimal_risky_asset_allocation.R
\name{calc_optimal_risky_asset_allocation}
\alias{calc_optimal_risky_asset_allocation}
\title{Calculate optimal risky asset allocation}
\usage{
calc_optimal_risky_asset_allocation(
  risky_asset_return_mean,
  risky_asset_return_sd,
  safe_asset_return,
  risk_aversion
)
}
\arguments{
\item{risky_asset_return_mean}{A numeric. The expected return of the risky asset.}

\item{risky_asset_return_sd}{A numeric. The standard deviation of the returns of the risky asset.}

\item{safe_asset_return}{A numeric. The expected return of the safe asset.}

\item{risk_aversion}{A numeric. The risk aversion coefficient.}
}
\description{
Calculate optimal risky asset allocation
}
