% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_purchasing_power.R
\name{calc_purchasing_power}
\alias{calc_purchasing_power}
\title{Calculate purchasing power}
\usage{
calc_purchasing_power(x, years, real_interest_rate)
}
\arguments{
\item{x}{A numeric. The initial amount of money.}

\item{years}{A numeric. The number of years.}

\item{real_interest_rate}{A numeric. The yearly avarage real interest rate.
The real interest rate is the interest rate after inflation.
If negative (e.g. equal to the avarage yearly inflation rate)
it can show deminishing purchasing power over time.}
}
\value{
A numeric. The purchasing power.
}
\description{
Calculate purchasing power
}
