% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_risk_adjusted_return.R
\name{calc_risk_adjusted_return}
\alias{calc_risk_adjusted_return}
\title{Calculate risk adjusted return}
\usage{
calc_risk_adjusted_return(
  safe_asset_return,
  risky_asset_return_mean,
  risky_asset_allocation,
  risky_asset_return_sd = NULL,
  risk_aversion = NULL
)
}
\arguments{
\item{safe_asset_return}{A numeric. The expected return of the safe asset.}

\item{risky_asset_return_mean}{A numeric. The expected return of the risky asset.}

\item{risky_asset_allocation}{A numeric. The allocation to the risky asset.
If it is already optimal allocation then parameters
\code{risky_asset_return_sd} and \code{risk_aversion} can be omitted.}

\item{risky_asset_return_sd}{A numeric. The standard deviation of the returns of the risky asset.}

\item{risk_aversion}{A numeric. The risk aversion coefficient.}
}
\description{
Calculate risk adjusted return
}
