% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_gompertz.R
\name{plot_joint_survival}
\alias{plot_joint_survival}
\title{Plotting the results of Gompertz model calibration for joint survival}
\usage{
plot_joint_survival(params, include_gompertz = FALSE)
}
\arguments{
\item{params}{A list returned by \code{\link[=calc_gompertz_joint_parameters]{calc_gompertz_joint_parameters()}} function.}

\item{include_gompertz}{A logical. Should the Gompertz survival curve be included in the plot?}
}
\description{
Plotting the results of Gompertz model calibration for joint survival
}
