% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_purchasing_power.R
\name{plot_purchasing_power}
\alias{plot_purchasing_power}
\title{Plotting changes to purchasing power over time}
\usage{
plot_purchasing_power(
  x,
  real_interest_rate,
  years = 50,
  legend_title = "Real interest rate",
  seed = NA
)
}
\arguments{
\item{x}{A numeric. The initial amount of money.}

\item{real_interest_rate}{A numeric. The yearly avarage real interest rate.
The real interest rate is the interest rate after inflation.
If negative (e.g. equal to the avarage yearly inflation rate)
it can show deminishing purchasing power over time.}

\item{years}{A numeric. The number of years.}

\item{legend_title}{A character. Could be "Inflation rate" if real interest
rate is negative.}

\item{seed}{A numeric. Seed passed to \code{geom_label_repel()}.}
}
\description{
Plotting changes to purchasing power over time
}
