% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_retirement_ruin.R
\name{plot_retirement_ruin}
\alias{plot_retirement_ruin}
\title{Plotting retirement ruin}
\usage{
plot_retirement_ruin(
  portfolio_return_mean,
  portfolio_return_sd,
  age,
  gompertz_mode,
  gompertz_dispersion,
  portfolio_value,
  monthly_spendings = NULL
)
}
\arguments{
\item{portfolio_return_mean}{A numeric. Mean of portfolio returns.}

\item{portfolio_return_sd}{A numeric. Standard deviation of portfolio returns.}

\item{age}{A numeric. Current age.}

\item{gompertz_mode}{A numeric. Gompertz mode.}

\item{gompertz_dispersion}{A numeric. Gompertz dispersion.}

\item{portfolio_value}{A numeric. Initial portfolio value.}

\item{monthly_spendings}{A numeric. Monthly spendings.}
}
\description{
Plotting retirement ruin
}
