% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_risk_adjusted_returns.R
\name{plot_risk_adjusted_returns}
\alias{plot_risk_adjusted_returns}
\title{Plotting risk adjusted returns}
\usage{
plot_risk_adjusted_returns(
  safe_asset_return,
  risky_asset_return_mean,
  risky_asset_return_sd,
  risk_aversion = 2,
  current_risky_asset_allocation = NULL
)
}
\arguments{
\item{safe_asset_return}{A numeric. The expected return of the safe asset.}

\item{risky_asset_return_mean}{A numeric. The expected return of the risky asset.}

\item{risky_asset_return_sd}{A numeric. The standard deviation of the returns of the risky asset.}

\item{risk_aversion}{A numeric. The risk aversion coefficient.}

\item{current_risky_asset_allocation}{A numeric.
The current allocation to the risky asset.}
}
\description{
Plotting risk adjusted returns
}
