% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R4GoodPersonalFinances-package.R
\docType{package}
\name{R4GoodPersonalFinances-package}
\alias{R4GoodPersonalFinances}
\alias{R4GoodPersonalFinances-package}
\title{R4GoodPersonalFinances: Make Optimal Financial Decisions}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Make optimal decisions for your personal or household finances. Use tools and methods that are selected carefully to align with academic consensus, bridging the gap between theoretical knowledge and practical application. They help you find your own personalized optimal discretionary spending or optimal asset allocation, and prepare you for retirement or financial independence. The optimal solution to this problems is extremely complex, and we only have a single lifetime to get it right. Fortunately, we now have the user-friendly tools implemented, that integrate life-cycle models with single-period net-worth mean-variance optimization models. Those tools can be used by anyone who wants to see what highly-personalized optimal decisions can look like. For more details see: Idzorek T., Kaplan P. (2024, ISBN:9781952927379), Haghani V., White J. (2023, ISBN:9781119747918).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.r4good.academy/}
  \item \url{https://r4goodacademy.github.io/R4GoodPersonalFinances/}
  \item \url{https://github.com/R4GoodAcademy/R4GoodPersonalFinances}
  \item Report bugs at \url{https://github.com/R4GoodAcademy/R4GoodPersonalFinances/issues}
}

}
\author{
\strong{Maintainer}: Kamil Wais \email{kamil.wais@gmail.com} (\href{https://orcid.org/0000-0002-4062-055X}{ORCID}) [copyright holder, funder]

Authors:
\itemize{
  \item Olesia Wais \email{olesia.wais@gmail.com} (\href{https://orcid.org/0000-0002-8741-8674}{ORCID})
}

}
