% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run a package app}
\usage{
run_app(which = "optimal-allocation")
}
\arguments{
\item{which}{A character. The name of the app to run.
Currently available:
\itemize{
\item \code{optimal-allocation} - Optimal Risky Asset Allocation
}}
}
\description{
Run a package app
}
