% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run a package app}
\usage{
run_app(
  which = c("risk-adjusted-returns", "purchasing-power"),
  res = 120,
  shinylive = FALSE
)
}
\arguments{
\item{which}{A character. The name of the app to run.
Currently available:
\itemize{
\item \code{risk-adjusted-returns} - Plotting risk-adjusted returns for various allocations to the risky asset allows you to find the optimal allocation.
\item \code{purchasing-power} - Plotting the effect of real interest rates
(positive or negative) on the purchasing power of savings over time.
}}

\item{res}{A numeric. The initial resolution of the plots.}

\item{shinylive}{A logical. Whether to use \code{shinylive} for the app.}
}
\description{
Run a package app
}
